server a funcionar com menos classes
